backend/apps/analytics/views.py
"""
Analytics Views
"""
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework.permissions import IsAuthenticated


class DashboardAnalyticsView(APIView):
    """Get dashboard analytics"""
    permission_classes = [IsAuthenticated]
    
    def get(self, request):
        return Response({
            'message': 'Dashboard Analytics API',
            'data': {}
        })


class LearningMetricsView(APIView):
    """Get learning metrics"""
    permission_classes = [IsAuthenticated]
    
    def get(self, request):
        return Response({
            'message': 'Learning Metrics API',
            'metrics': {}
        })


class TopicBreakdownView(APIView):
    """Get topic breakdown"""
    permission_classes = [IsAuthenticated]
    
    def get(self, request):
        return Response({
            'message': 'Topic Breakdown API',
            'topics': []
        })


class TrendsView(APIView):
    """Get learning trends"""
    permission_classes = [IsAuthenticated]
    
    def get(self, request):
        return Response({
            'message': 'Trends API',
            'trends': []
        })
