backend/requirements.txt
# Backend Core Dependencies
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
django-filter>=24.0
django-redis>=5.4.0
djangorestframework-simplejwt>=5.3.0
django-storages>=1.14.0
django-celery-beat>=2.6.0
django-celery-results>=2.5.0

# FastAPI (Hybrid API)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
psycopg[binary]>=3.1.0
pgvector>=0.2.4
neo4j>=5.16.0
redis>=5.0.0

# Async & Task Queue
celery>=5.3.0
channels>=4.0.0
channels-redis>=4.2.0
kafka-python>=2.0.2

# ML/AI - PyTorch
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
pytorch-lightning>=2.2.0

# ML/AI - Transformers & NLP
transformers>=4.38.0
sentence-transformers>=2.4.0
tokenizers>=0.15.0
datasets>=2.17.0
accelerate>=0.27.0

# ML/AI - Vision
timm>=0.9.12
opencv-python>=4.9.0
Pillow>=10.2.0
albumentations>=1.3.1

# ML/AI - Audio/Video
whisper>=1.1.10
librosa>=0.10.1
moviepy>=1.0.3

# ML/AI - Reinforcement Learning
stable-baselines3>=2.2.0
gymnasium>=0.29.1

# ML/AI - Graph Neural Networks
torch-geometric>=2.5.0
networkx>=3.2.0

# ML/AI - Spiking Neural Networks
snnTorch>=0.8.0
norse>=1.0.0

# Document Processing
pdfplumber>=0.10.0
python-docx>=1.1.0
reportlab>=4.1.0
pymupdf>=1.23.0

# Report Generation
WeasyPrint>=60.2
Jinja2>=3.1.3
markdown>=3.5.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0
aiohttp>=3.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
boto3>=1.34.0
structlog>=24.1.0

# Monitoring & Profiling
prometheus-client>=0.19.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-instrumentation-django>=0.44b0

# Testing
pytest>=8.0.0
pytest-django>=4.8.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=23.0.0

# Development
black>=24.2.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0
